docker_compose/simple_project/app/requirements/base.txt
Django==5.0.4
psycopg[binary]==3.1.18
python-dotenv==1.0.1
split-settings==1.0.0